+++ src/Http/Resources/GroupResource.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Namu\WireChat\Http\Resources;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class GroupResource extends JsonResource
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @return array<string, mixed>
+     */
+    public function toArray(Request $request): array
+    {
+         return [
+            'id' => $this->id,
+            'conversation_id' => $this->conversation_id,
+            'name' => $this->name,
+            'cover_url' =>$this->cover_url
+        ];
+    }
+}